improving the tests

+++ alyanovikova/PersistenceTask/Tests/PersistenceTaskTests/PersistenceTaskTests.swift
@@ -3,23 +3,20 @@ import XCTest
 
 final class PersistenceTaskTests: XCTestCase {
   private var fileURL: URL!
-  
-  override func setUp() {
-    super.setUp()
-    do {
-      let documentsDirectory = try FileManager.default.url(
-        for: .documentDirectory,
-        in: .userDomainMask,
-        appropriateFor: nil,
-        create: true)
-      fileURL = documentsDirectory.appendingPathComponent("TodoList")
-    } catch {
-      XCTFail("Test failed with \(error)")
+
+  override func setUpWithError() throws {
+    fileURL = try TodoList.makeDefaultURL()
+    if FileManager.default.fileExists(atPath: fileURL.path) {
+      try FileManager.default.removeItem(at: fileURL)
     }
   }
+
+  func testInitialization() throws {
+     _ = try TodoList()
+  }
   
-  func testReadingFromFile() throws {
-    try FileManager.default.removeItem(at: fileURL)
+  func testNotEmptyInitialization() throws {
+    // GIVEN
     do {
       let todos = [Todo(todo: "take a walk with dogs"), Todo(todo: "learn swift"), Todo(todo: "call mum")]
       
@@ -30,9 +27,11 @@ final class PersistenceTaskTests: XCTestCase {
     } catch {
       XCTFail("Test failed with \(error)")
     }
-    
+
+    // WHEN
     let todoList = try TodoList()
-    
+
+    // THEN
     XCTAssert(FileManager.default.fileExists(atPath: fileURL.path))
     XCTAssert(todoList.todos.count == 3)
     XCTAssert(todoList.todos[0].todo == "take a walk with dogs" && todoList.todos[0].isCompleted == false)
@@ -40,8 +39,8 @@ final class PersistenceTaskTests: XCTestCase {
     XCTAssert(todoList.todos[2].todo == "call mum" && todoList.todos[0].isCompleted == false)
   }
   
-  func testWritingToFile() throws {
-    try FileManager.default.removeItem(at: fileURL)
+  func testAddAndChangeStatus() throws {
+    // WHEN
     do {
       let todoList = try TodoList()
       try todoList.add(todo: Todo(todo: "get some sleep"))
@@ -50,7 +49,8 @@ final class PersistenceTaskTests: XCTestCase {
     } catch {
       XCTFail("Test failed with \(error)")
     }
-    
+
+    // THEN
     let data = try Data(contentsOf: fileURL)
     
     let jsonDecoder = JSONDecoder()
@@ -60,9 +60,4 @@ final class PersistenceTaskTests: XCTestCase {
     XCTAssert(todos[0].todo == "get some sleep" && todos[0].isCompleted == true)
     XCTAssert(todos[1].todo == "buy curd snack" && todos[1].isCompleted == false)
   }
-  
-  static var allTests = [
-    ("testWritingToFile", testWritingToFile),
-    ("testReadingFromFile", testReadingFromFile),
-  ]
 }
